+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- getorganized: Added functions and tests for using GetOrganized.
+
 ## [2.9.0] - 2025-02-19
 
 ### Added